pyproject.toml
[project]
name = "ttrpg-listen"
version = "0.3.0"
description = "Local AI-powered live transcription for TTRPG sessions"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "pyaudiowpatch>=0.2.12",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "scipy>=1.11",
    "whisperx>=3.8",
    "pyannote-audio>=3.1",
    "pandas",
    "comtypes",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[project.scripts]
ttrpglisten = "ttrpglisten.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
